+++ src/main/java/com/github/lemniscate/spring/crud/web/ApiResourceController.java
@@ -4,6 +4,7 @@ import com.github.lemniscate.spring.crud.mapping.ApiResourceMapping;
 import com.github.lemniscate.spring.crud.svc.ApiResourceService;
 import com.github.lemniscate.spring.crud.util.ApiResourceUtil;
 import com.github.lemniscate.spring.crud.web.assembler.ApiResourceAssembler;
+import com.github.lemniscate.spring.crud.web.assembler.ApiResourceAssemblers;
 import lombok.Getter;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.data.domain.Page;
@@ -27,13 +28,14 @@ import java.util.Map;
 */
 public class ApiResourceController<ID extends Serializable, E extends Identifiable<ID>, CB, RB extends Identifiable<ID>, UB> {
 
-    public static final String X_SELF_HREF = "X-SELF-HREF";
-
     @Inject
     protected ApiResourceService<ID, E, CB, RB, UB> service;
 
     @Inject
     protected ApiResourceAssembler<ID, E, CB, RB, UB> assembler;
+
+    @Inject
+    protected ApiResourceAssemblers assemblers;
 
     @Getter
     protected final ApiResourceMapping<ID, E, CB, RB, UB> mapping;
@@ -50,25 +52,24 @@ public class ApiResourceController<ID extends Serializable, E extends Identifiab
         this((ApiResourceMapping<ID, E, CB, RB, UB>) ApiResourceUtil.generateMapping(3, ApiResourceController.class));
     }
 
+    public ApiResourceMapping<ID, E, CB, RB, UB> getMapping() {
+        return mapping;
+    }
+
     @RequestMapping(value="", method= RequestMethod.GET)
     public ResponseEntity<Page<Resource<RB>>> getAll(@RequestParam MultiValueMap<String, String> params, Pageable p){
         Page<RB> entities = service.findForRead(p);
-        List<Resource<RB>> resources = assembler.toResources(entities);
-        Page<Resource<RB>> pagedResources = new PageImpl<Resource<RB>>(resources, p, entities.getTotalElements());
-        ResponseEntity<Page<Resource<RB>>> response = new ResponseEntity<Page<Resource<RB>>>(pagedResources, HttpStatus.OK);
-        return response;
+        return assemblers.respond(mapping.readBeanClass(), entities, p, HttpStatus.OK);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.GET)
     public ResponseEntity<Resource<RB>> getOne(@PathVariable ID id){
         RB entity = service.read(conversionService.convert(id, mapping.idClass()));
-        Resource<RB> resource = assembler.toResource(entity);
-        ResponseEntity<Resource<RB>> response = new ResponseEntity<Resource<RB>>(resource, HttpStatus.OK);
-        return response;
+        return assemblers.respond(entity, entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-    public ResponseEntity<Void> deleteOne(@PathVariable ID id){
+    public ResponseEntity<?> deleteOne(@PathVariable ID id){
         service.delete(conversionService.convert(id, mapping.idClass()));
         return new ResponseEntity<Void>(HttpStatus.OK);
     }
@@ -76,34 +77,19 @@ public class ApiResourceController<ID extends Serializable, E extends Identifiab
     @RequestMapping(value="", method=RequestMethod.POST)
     public ResponseEntity<Resource<RB>> postOne(@RequestBody CB bean){
         RB entity = service.createForRead(bean);
-        Resource<RB> resource = assembler.toResource(entity);
-        MultiValueMap<String,String> headers = new LinkedMultiValueMap<String, String>();
-        headers.add(X_SELF_HREF, resource.getLink("self").getHref() );
-        ResponseEntity<Resource<RB>> response = new ResponseEntity<Resource<RB>>(resource, headers, HttpStatus.CREATED);
-        return response;
+        return assemblers.respond(entity, HttpStatus.CREATED);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.PUT)
     public ResponseEntity<Resource<RB>> putOne(@PathVariable ID id, @RequestBody UB bean){
         RB entity = service.updateForRead( id, bean );
-        Resource<RB> resource = assembler.toResource(entity);
-        ResponseEntity<Resource<RB>> response = new ResponseEntity<Resource<RB>>(resource, HttpStatus.OK);
-        return response;
-    }
-
-    public ApiResourceMapping<ID, E, CB, RB, UB> getMapping() {
-        return mapping;
+        return assemblers.respond(entity, entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
     }
 
     @RequestMapping(value="/searches", method=RequestMethod.POST)
     public ResponseEntity<Page<Resource<RB>>> search(@RequestBody Map<String, Object> search, Pageable pageable){
         Page<RB> entities = service.searchForRead(search, pageable);
-        List<Resource<RB>> resources = assembler.toResources(entities.getContent());
-        Page<Resource<RB>> pagedResources = new PageImpl<Resource<RB>>(resources, pageable, entities.getTotalElements());
-        ResponseEntity<Page<Resource<RB>>> response = new ResponseEntity<Page<Resource<RB>>>(pagedResources, HttpStatus.OK);
-        return response;
+        return assemblers.respond( mapping.readBeanClass(), entities, pageable, HttpStatus.OK);
     }
-
-
 
 }